requirements.txt
# ============================================
# REQUIREMENTS.TXT
# Unreal AI Game Development Platform
# Complete Dependencies for Python 3.11+
# ============================================
# Last Updated: February 17, 2026
# Tested and Production-Ready

# ============================================
# PYTHON VERSION REQUIREMENT
# ============================================
# Python>=3.11,<4.0

# ============================================
# CORE FRAMEWORK & WEB
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# ============================================
# AI & MACHINE LEARNING
# ============================================
openai==1.10.0
anthropic==0.8.1
transformers==4.37.0
torch>=2.1.0
torchvision>=0.16.0

# ============================================
# ASYNC & NETWORKING
# ============================================
aiohttp==3.9.1
aiofiles==23.2.1
httpx==0.26.0
requests==2.31.0
aiosmtplib==3.0.1

# ============================================
# DATABASE & CACHING
# ============================================
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
hiredis==2.3.2
cachetools==5.3.2
diskcache==5.6.3

# ============================================
# AUTHENTICATION & SECURITY
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
authlib==1.3.0
cryptography==42.0.0

# ============================================
# FILE & MEDIA PROCESSING
# ============================================
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
ffmpeg-python==0.2.0
imageio==2.33.1
imageio-ffmpeg==0.4.9
# ============================================
# DOCUMENT PROCESSING
# ============================================
pypdf2==3.0.1
reportlab==4.0.9
openpyxl==3.1.2
xlsxwriter==3.1.9
pandas==2.2.0
python-docx==1.1.0
python-pptx==0.6.23

# ============================================
# TEMPLATING & CODE GENERATION
# ============================================
jinja2==3.1.3
markupsafe==2.1.4

# ============================================
# AUDIO & MUSIC PROCESSING
# ============================================
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
music21==9.1.0

# ============================================
# NATURAL LANGUAGE PROCESSING
# ============================================
nltk==3.8.1
spacy==3.7.2

# ============================================
# DATA SCIENCE & ANALYTICS
# ============================================
scipy==1.12.0
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.2

# ============================================
# 3D & GEOMETRY PROCESSING
# ============================================
trimesh==4.0.10
pygltflib==1.16.1

# ============================================
# CLOUD & DEPLOYMENT (Azure)
# ============================================
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-monitor-opentelemetry==1.2.0
opencensus-ext-azure==1.1.13

# ============================================
# MONITORING & LOGGING
# ============================================
prometheus-client==0.19.0
sentry-sdk==1.40.0
python-json-logger==2.0.7
loguru==0.7.2

# ============================================
# RATE LIMITING & MIDDLEWARE
# ============================================
slowapi==0.1.9

# ============================================
# TASK QUEUE & BACKGROUND JOBS
# ============================================
celery==5.3.6
kombu==5.3.5

# ============================================
# TESTING & CODE QUALITY
# ============================================
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
black==24.1.1
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0

# ============================================
# UTILITIES & HELPERS
# ============================================
python-dotenv==1.0.1
python-dateutil==2.8.2
pytz==2024.1
marshmallow==3.20.2
click==8.1.7
typer==0.9.0
tqdm==4.66.1
pyyaml==6.0.1
orjson==3.9.13
urllib3==2.2.0
toml==0.10.2
configparser==6.0.0

# ============================================
# WEBSOCKETS & REAL-TIME
# ============================================
python-socketio==5.11.1

# ============================================
# VERSION CONTROL
# ============================================
gitpython==3.1.41

# ============================================
# CONTAINERIZATION
# ============================================
docker==7.0.0

# ============================================
# OPTIONAL: PAYMENT PROCESSING
# ============================================
stripe==8.1.0

# ============================================
# OPTIONAL: SOCIAL AUTHENTICATION
# ============================================
PyGithub==2.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# ============================================
# SEARCH & INDEXING
# ============================================
whoosh==2.7.4
elasticsearch==8.12.0

# ============================================
# DEVELOPMENT & DEBUGGING
# ============================================
watchdog==4.0.0
ipdb==0.13.13
ipython==8.21.0

# ============================================
# DOCUMENTATION
# ============================================
mkdocs==1.5.3
mkdocs-material==9.5.7

# ============================================
# PRODUCTION DEPLOYMENT
# ============================================
gunicorn==21.2.0
supervisor==4.2.5

# ============================================
# PLATFORM-SPECIFIC PACKAGES
# ============================================
pywin32==306; sys_platform == 'win32'
pyobjc==10.1; sys_platform == 'darwin'